move validations logics to a service
